Use wrap mode by default
