+++ packages/transforms/hoist-field/package.json
@@ -23,11 +23,12 @@
     "graphql": "*"
   },
   "dependencies": {
-    "@graphql-mesh/types": "0.49.0",
-    "@graphql-mesh/utils": "0.16.2",
-    "@graphql-tools/delegate": "8.2.1",
-    "@graphql-tools/wrap": "8.1.0",
-    "@graphql-tools/utils": "8.2.2"
+    "@graphql-mesh/types": "0.68.1",
+    "@graphql-mesh/utils": "0.31.0",
+    "@graphql-tools/delegate": "8.5.4",
+    "@graphql-tools/wrap": "8.4.6",
+    "@graphql-tools/utils": "8.6.3",
+    "@graphql-mesh/cache-inmemory-lru": "0.6.3"
   },
   "publishConfig": {
     "access": "public",
