--- packages/transforms/hoist-field/src/types.ts
@@ -1,3 +1,0 @@
-import { YamlConfig } from '@graphql-mesh/types';
-
-export type HoistFieldTransformFieldPathConfig = YamlConfig.HoistFieldTransformFieldConfigObject['pathConfig'][0];
